refactor(API): Updated the configurations accepted by Circe
We changed the configuration of Circe to accept a new configuration pluginCfgs that can be passed
at any stage including the subscribe and publish methods. These options will be passed down to the
child plugin for it's use.

BREAKING CHANGE: publishCfgs,consumeCfgs,producerCfgs and consumerCfgs are all pluginCfgs at all
method calls.