feat(ProducerAPI): Refactored Producer API
Added middleware capabailities that allow for pre and post validation of messages as well as a key
generator that generates a uuid v4 key and can also attach it to a message object if the keyProp
option is defined. Also added tests.

Validators should no longer return anything. They strictly allow for users to validate a message and
handle the validation in whatever way they see fit. If they don't want to process a message, they
should throw from inside the validator.