feat(API): Consumer can handle multiple events/handlers
A consumer can now handle multiple events and multiple event handlers if needed. It can also just
take a topic and a handler. I also added Flow types and now validates to make sure that a message
received from kafka as a buffer can be parsed from the string(that it's JSON) otherwise it will
return the string.

Consumer now has a method called subscribe to make it less forceful since we know accept multiple
events.