feat(API): Added a test for API exposure which is just exporting the create modules from producer/co
Added a test for the module exposing the create methods for producer and consumers. This includes a
basic test for this. Updated middleware flow types to be more descriptive. Updated packages and
lock.json and added a LICENSE under Apache 2.0

This is no longer a factory function and instead just exposes individual create methods for
producers and consumers