refactor(createConsumer): CreateConsumer subscribe method now is separate
Before, the createConsumer returned an object with a subscribe method that when called would both
subscribe to a topic and set up the handler. This led to problems with being able to unsubscribe(but
keep the same handler) Now the subscribe method does not start consuming. A new consume method takes
the handler and middleware now.

subscribe no longer starts consuming, consume must be called before a handler is configured.